requirements.txt
Flask==3.0.3
tensorflow==2.17.0
tf-keras==2.17.0
transformers==4.45.1
matplotlib==3.8.3
markdown==3.7
google-generativeai==0.8.2
numpy==1.26.4
